add balanceTransaction func
this function identifies any tx posting amount that needs to be inferred
and sets its value, while confirming the transaction is balanced.

sanity test was updated to compare expected to parsed tx, which includes
a posting that must be inferred. benches were updated and ran to include
this new operation

all tests pass