refactor types
This changes how we represent commodities. They are now stored in Lots, each lot having a
date, amount, unit value, and commodity type & code. When parsing commodities, we determine
and store its code from the format instead of storing the format. A default currency is used in
cases where no code is provided.

A list of known/supported currencies is maintained.
Any commodity not identified as a currency is assumed to be a stock/security with a unit value.

All tests passing.