Reload recent transactions list when journal file changes
When a new file is selected, or when new transactions are added to existing file, then
the recent transactions list needs to be reloaded/rendered.  This is done using htmx
out of bound update (hx-swap-oob), which means we need to return an element to
replace the recent-tx list in the responses to selecting a file and adding an expense.

This is done in a round-about way by just returning an empty element with instructions
to get recent-tx on load. That empty element needs to be defined in the add-expense
and file-selector templates. Not ideal, but works for now.